image-gen/requirements.txt
openai==1.40.0
langchain==0.2.15
langchain-openai==0.1.20
python-dotenv==1.0.0
pillow==10.4.0
requests==2.32.3

